update shader library to only build required shaders and shaders for materials that have been imported as modules
